feat: refactor password rules import path; remove unused loading components and hooks
