+++ src/app/(auth)/register/components/register-password-fields.tsx
@@ -4,7 +4,7 @@ import { useState } from "react"
 import { Eye, EyeOff } from "lucide-react"
 import { Input } from "@/components/ui/input"
 import { Label } from "@/components/ui/label"
-import { PasswordRules } from "@/components/password-rules"
+import { PasswordRules } from "@/components/password-rules/password-rules"
 import { usePasswordRules } from "@/hooks/use-password-rules"
 import { INVALID_INPUT_CLASS } from "@/lib/form-styles"
 import { cn } from "@/lib/utils"
