+++ src/app/(auth)/reset-password/reset-password-form.tsx
@@ -7,7 +7,7 @@ import { AppRoutes } from "@/constants/routes"
 import { Button } from "@/components/ui/button"
 import { Input } from "@/components/ui/input"
 import { Label } from "@/components/ui/label"
-import { PasswordRules } from "@/components/password-rules"
+import { PasswordRules } from "@/components/password-rules/password-rules"
 import { usePasswordRules } from "@/hooks/use-password-rules"
 import { AuthFormShell } from "@/components/auth/auth-form-shell"
 import { INVALID_INPUT_CLASS } from "@/lib/form-styles"
