--- src/components/loading-page.tsx
@@ -1,9 +1,0 @@
-import { Loader2 } from "lucide-react"
-
-export function LoadingPage() {
-  return (
-    <div className="min-h-screen flex items-center justify-center">
-      <Loader2 className="h-8 w-8 animate-spin text-muted-foreground"/>
-    </div>
-  )
-}
