--- src/components/loading/loading.tsx
@@ -1,7 +1,0 @@
-import { Loader2 } from 'lucide-react'
-
-export function Loading() {
-  return (
-    <Loader2 className="h-8 w-8 animate-spin text-muted-foreground"/>
-  )
-}
